refactor: create separate files for root and schema
